Added AdminBadgeCommand

+++ src/main/java/de/skymyth/badge/BadgeManager.java
@@ -3,12 +3,15 @@ package de.skymyth.badge;
 import de.skymyth.SkyMythPlugin;
 import de.skymyth.badge.model.Badge;
 import de.skymyth.badge.repository.BadgeRepository;
+import eu.decentsoftware.holograms.api.utils.scheduler.S;
 import lombok.Getter;
 import lombok.experimental.FieldDefaults;
 
 import java.util.ArrayList;
 import java.util.List;
+import java.util.Map;
 import java.util.UUID;
+import java.util.stream.Collectors;
 
 @Getter
 @FieldDefaults(level = lombok.AccessLevel.PRIVATE, makeFinal = true)
@@ -17,31 +20,42 @@ public class BadgeManager {
     SkyMythPlugin plugin;
     BadgeRepository badgeRepository;
 
-    List<Badge> cachedBadges;
+    Map<String, Badge> cachedBadges;
 
     public BadgeManager(SkyMythPlugin plugin) {
         this.plugin = plugin;
         this.badgeRepository = plugin.getMongoManager().create(BadgeRepository.class);
 
-        this.cachedBadges = new ArrayList<>(badgeRepository.findAll());
+        this.cachedBadges = badgeRepository.findAll().stream().collect(Collectors.toMap(Badge::getName, badge -> badge));
     }
 
     public void createBadge(Badge badge) {
+        if (cachedBadges.values().stream().anyMatch(b -> b.getName().equalsIgnoreCase(badge.getName()))) {
+            throw new IllegalArgumentException("Badge with name " + badge.getName() + " already exists.");
+        }
         badgeRepository.save(badge);
-        cachedBadges.add(badge);
+        cachedBadges.put(badge.getName(), badge);
     }
 
     public void deleteBadge(Badge badge) {
+        if (!cachedBadges.containsValue(badge)) {
+            throw new IllegalArgumentException("Badge with name " + badge.getName() + " does not exist.");
+        }
         badgeRepository.delete(badge);
-        cachedBadges.remove(badge);
+        cachedBadges.remove(badge.getName());
+    }
+
+    public void saveBadge(Badge badge) {
+        cachedBadges.put(badge.getName(), badge);
+        badgeRepository.save(badge);
     }
 
     public Badge getBadge(String name) {
-        return cachedBadges.stream().filter(badge -> badge.getName().equalsIgnoreCase(name)).findFirst().orElse(null);
+        return cachedBadges.values().stream().filter(badge -> badge.getName().equalsIgnoreCase(name)).findFirst().orElse(null);
     }
 
     public List<Badge> getBadges(UUID user) {
-        return cachedBadges.stream().filter(badge -> badge.getOwners().contains(user)).toList();
+        return cachedBadges.values().stream().filter(badge -> badge.getOwners().contains(user)).toList();
     }
 
 }
